--- src/pages/mission/missonCard/MissionSubContent.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const MissionSubContentCategory = ({}) => {
-    return <div></div>;
-};
-
-export default MissionSubContentCategory;
