+++ src/stories/pages/mission/MissionContent.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react';
 import MissionContent, {
     IMissionContentProps,
-} from '@/pages/mission/missonCard/MissionContent';
+} from '@/pages/mission/missionCard/MissionContent';
 
 const meta: Meta<IMissionContentProps> = {
     title: 'mission/MissionContent',
@@ -13,21 +13,6 @@ const meta: Meta<IMissionContentProps> = {
                 type: 'text',
             },
             description: '미션 이름',
-        },
-        category: {
-            control: {
-                type: 'text',
-            },
-            description: '카테고리',
-        },
-        difficulty: {
-            control: {
-                type: 'range',
-                min: 0,
-                max: 3,
-                step: 1,
-            },
-            description: '난이도',
         },
     },
     parameters: {
@@ -42,7 +27,5 @@ type Story = StoryObj<typeof meta>;
 export const Primary: Story = {
     args: {
         name: '미션 이름',
-        category: '카테고리',
-        difficulty: 3,
     },
 };
